create code cell component to have multiple editors
